Create commit2
We wrote a code that checks if any of the algorithms are used. For example, if Algorithm 1 is used, it returns the corresponding value X1, if algorithm 3 is used it gives X3 value which is used and the code then calculates it in Java. If no algorithm is selected, the code will not perform any calculations. We still need to write the code which checks if any of the algorithm is used